+++ UWP/CSharp/Content/Shaders/VPRTVertexShader.hlsl
@@ -1,11 +1,11 @@
 // Per-vertex data passed to the geometry shader.
 struct VertexShaderOutput
 {
-    min16float4 pos     : SV_POSITION;
-    min16float3 color   : COLOR0;
+    float4 pos      : SV_POSITION;
+    float3 color    : COLOR0;
 
     // The render target array index is set here in the vertex shader.
-    uint        viewId  : SV_RenderTargetArrayIndex;
+    uint   viewId   : SV_RenderTargetArrayIndex;
 };
 
 #include "VertexShaderShared.hlsl"
